Fixed #79: Upgraded to PA SDK 2.0.0
